refactor(hub): extract terminal subsystem into terminal.go
Pure code move — no behavior change. Same package, same globals
(db, jwtSecret, agents/agentsMu, termSessions/termSessionsMu, rateLimiter).

hub/main.go: 3357 → 2904 lines, hub/terminal.go: 466 lines.

Moved (still package main):
  - TerminalSession type
  - maxTerminalSessions const
  - configuredAllowedOrigins (only used by terminal WS upgrader)
  - generateTerminalBrowserToken
  - handleStartTerminal
  - handleCloseTerminal
  - handleTerminalWS
  - waitForSessionEnd
  - terminalRelay
  - cleanupTerminalSession

go test -count=1 ./...  green 3 runs in a row.
